+++ sample/sample15-function-inject/PostManager.ts
@@ -1,0 +1,7 @@
+import {Service} from "../../src/decorators/Service";
+
+export const PostManager = Service(() => ({
+    getId() {
+        return "some post id";
+    }
+}));