+++ sample/sample15-function-inject/app.ts
@@ -1,0 +1,5 @@
+import {Container} from "../../src";
+import {PostController} from "./PostController";
+
+const postController = Container.get(PostController);
+console.log(postController);