+++ src/ContainerInstance.ts
@@ -1,10 +1,10 @@
-import { ServiceMetadata } from "./types/ServiceMetadata";
-import { ObjectType } from "./types/ObjectType";
-import { Token } from "./Token";
-import { ServiceIdentifier } from "./types/ServiceIdentifier";
-import { ServiceNotFoundError } from "./error/ServiceNotFoundError";
-import { MissingProvidedServiceTypeError } from "./error/MissingProvidedServiceTypeError";
-import { Container } from "./Container";
+import {Container} from "./Container";
+import {MissingProvidedServiceTypeError} from "./error/MissingProvidedServiceTypeError";
+import {ServiceNotFoundError} from "./error/ServiceNotFoundError";
+import {Token} from "./Token";
+import {ObjectType} from "./types/ObjectType";
+import {ServiceIdentifier} from "./types/ServiceIdentifier";
+import {ServiceMetadata} from "./types/ServiceMetadata";
 
 /**
  * TypeDI can have multiple containers.
@@ -90,7 +90,13 @@ export class ContainerInstance {
      * Retrieves the service with given name or type from the service container.
      * Optionally, parameters can be passed in case if instance is initialized in the container for the first time.
      */
-    get<T>(identifier: ServiceIdentifier): T {
+    get<T>(id: { service: T }): T;
+
+    /**
+     * Retrieves the service with given name or type from the service container.
+     * Optionally, parameters can be passed in case if instance is initialized in the container for the first time.
+     */
+    get<T>(identifier: ServiceIdentifier<T>): T {
 
         const globalContainer = Container.of(undefined);
         let service = globalContainer.findService(identifier);
@@ -174,12 +180,15 @@ export class ContainerInstance {
         if (typeof identifierOrServiceMetadata === "string" || identifierOrServiceMetadata instanceof Token) {
             return this.set({ id: identifierOrServiceMetadata, value: value });
         }
+        if (typeof identifierOrServiceMetadata === "object" && (identifierOrServiceMetadata as { service: Token<any> }).service) {
+            return this.set({ id: (identifierOrServiceMetadata as { service: Token<any> }).service, value: value });
+        }
         if (identifierOrServiceMetadata instanceof Function) {
             return this.set({ type: identifierOrServiceMetadata, id: identifierOrServiceMetadata, value: value });
         }
 
         // const newService: ServiceMetadata<any, any> = arguments.length === 1 && typeof identifierOrServiceMetadata === "object"  && !(identifierOrServiceMetadata instanceof Token) ? identifierOrServiceMetadata : undefined;
-        const newService: ServiceMetadata<any, any> = identifierOrServiceMetadata;
+        const newService: ServiceMetadata<any, any> = identifierOrServiceMetadata as any;
         const service = this.findService(newService.id);
         if (service && service.multiple !== true) {
             Object.assign(service, newService);
@@ -234,8 +243,15 @@ export class ContainerInstance {
      */
     private findService(identifier: ServiceIdentifier): ServiceMetadata<any, any> | undefined {
         return this.services.find(service => {
-            if (service.id)
+            if (service.id) {
+                if (identifier instanceof Object &&
+                    service.id instanceof Token &&
+                    (identifier as any).service instanceof Token) {
+                    return service.id === (identifier as any).service;
+                }
+
                 return service.id === identifier;
+            }
 
             if (service.type && identifier instanceof Function)
                 return service.type === identifier; // todo: not sure why it was here || identifier.prototype instanceof service.type;
@@ -270,6 +286,9 @@ export class ContainerInstance {
 
         } else if (identifier instanceof Function) {
             type = identifier;
+
+        // } else if (identifier instanceof Object && (identifier as { service: Token<any> }).service instanceof Token) {
+        //     type = (identifier as { service: Token<any> }).service;
         }
 
         // if service was not found then create a new one and register it
@@ -300,7 +319,7 @@ export class ContainerInstance {
                 value = (this.get(service.factory[0]) as any)[service.factory[1]](...params);
 
             } else { // regular factory function
-                value = service.factory(...params);
+                value = service.factory(...params, this);
             }
 
         } else {  // otherwise simply create a new object instance