+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.8.0
+
+* added new type of dependency injection - function DI
+* now null can be stored in the container for values
+
 ## 0.7.2
 
 * fixed bug with inherited services